refactor: 컴포넌트 명 변경
