+++ src/app/(private)/study-rooms/[id]/homework/[homeworkId]/edit/page.tsx
@@ -1,7 +1,7 @@
 import BackLink from '@/features/dashboard/studynote/components/back-link';
 import SelectArea from '@/features/dashboard/studynote/write/components/select-area';
-import { HomeworkFormProvider } from '@/features/homework/components/write/components/form-provider';
 import HomeworkEditForm from '@/features/homework/components/write/components/homework-edit-form';
+import { HomeworkFormProvider } from '@/features/homework/components/write/components/homework-form-provider';
 import { WriteArea } from '@/features/homework/components/write/components/write-area';
 import { ColumnLayout } from '@/layout/column-layout';
